backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.6

# AI/LLM
openai==1.3.9

# Testing
pytest==7.4.3
httpx==0.25.2

# WebSockets
websockets==12.0
